Workflows/requirements.txt
boto3>=1.28.0
langgraph>=0.0.15
streamlit>=1.31.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
botocore>=1.31.0